requirements.txt
langchain
langchain-ollama
ollama
sentence-transformers
faiss-cpu
PyMuPDF
spacy
requests
pyyaml
python-dotenv